Fixed  overlays in home screen
