+++ backend/src/auth/auth.interface.ts
@@ -1,6 +1,6 @@
 import { RegisterDto } from './dto/register.dto';
 import { LoginDto } from './dto/login.dto';
-import { AuthResponseDto } from './dto/auth-response.dto';
+import { AuthResponseDto, MessageResponseDto, RegisterResponseDto } from './dto/auth-response.dto';
 import { ForgotPasswordDto } from './dto/forgot-password.dto';
 import { ResetPasswordDto } from './dto/reset-password.dto';
 import { RefreshToken as PrismaRefreshToken, Prisma } from '@prisma/client';
